refactor: replace elements to dirs
