+++ frontend/src/pages/integrations/alice/components/smart-device/components/device-skills/capabilities/options-button/index.ts
@@ -1,0 +1,1 @@
+export { CapabilityOptionsButton } from './options-button';
