+++ frontend/src/pages/integrations/alice/components/smart-device/components/device-skills/capabilities/options-button/options-button.tsx
@@ -7,11 +7,8 @@ import {
   isFieldModified,
   type SmartDeviceCapability,
 } from '@/stores/alice';
-import { OptionsDivider } from '../options-divider';
-import { OptionsItem } from '../options-item';
-import { OptionsPopup } from '../options-popup';
-import { type CapabilitySubProps } from '../types';
-import { updateItem } from '../update-item';
+import { OptionsDivider, OptionsItem, OptionsPopup, updateItem } from '../../options';
+import { type CapabilitySubProps } from '../../types';
 
 export const CapabilityOptionsButton = ({
   capability, index, capabilities, onCapabilityChange,
