+++ frontend/src/pages/integrations/alice/components/smart-device/components/device-skills/options/options-item.tsx
