+++ frontend/src/pages/integrations/alice/components/smart-device/components/device-skills/options/options-popup.tsx
