+++ frontend/src/pages/integrations/alice/components/smart-device/components/device-skills/options/update-item.ts
