+++ frontend/src/pages/integrations/alice/components/smart-device/components/device-skills/property-options-button/index.ts
@@ -1,0 +1,1 @@
+export { PropertyOptionsButton } from './property-options-button';
